third_party/WebKit/LayoutTests/http/tests/inspector/service-workers/service-workers-view-expected.txt

CONSOLE MESSAGE: line 98: registerServiceWorker success
CONSOLE MESSAGE: line 98: registerServiceWorker success
CONSOLE MESSAGE: line 115: unregisterServiceWorker success
CONSOLE MESSAGE: line 115: unregisterServiceWorker success
Tests ServiceWorkersView on resources panel.

Select ServiceWorkers tree element.
Register ServiceWorker for scope1
==== ServiceWorkersView ====
127.0.0.1

Scope: /inspector/service-workers/resources/scope1/
activated
inspectrunning
Script: /inspector/service-workers/resources/service-worker-empty.js
LAST-MODIFIED
RESPONSE-TIME

============================
Register ServiceWorker for scope2
==== ServiceWorkersView ====
127.0.0.1

Scope: /inspector/service-workers/resources/scope1/
activated
inspectrunning
Script: /inspector/service-workers/resources/service-worker-empty.js
LAST-MODIFIED
RESPONSE-TIME

Scope: /inspector/service-workers/resources/scope2/
activated
inspectrunning
Script: /inspector/service-workers/resources/service-worker-empty.js
LAST-MODIFIED
RESPONSE-TIME

============================
Unegister ServiceWorker for scope1
==== ServiceWorkersView ====
127.0.0.1

Scope: /inspector/service-workers/resources/scope1/ - deleted
redundant
inspectrunning
Script: /inspector/service-workers/resources/service-worker-empty.js
LAST-MODIFIED
RESPONSE-TIME

Scope: /inspector/service-workers/resources/scope2/
activated
inspectrunning
Script: /inspector/service-workers/resources/service-worker-empty.js
LAST-MODIFIED
RESPONSE-TIME

============================
Unegister ServiceWorker for scope1
